requirements.txt
streamlit==1.29.0
langchain>=0.2.0
langchain-google-genai>=3.0.0
langchain-experimental>=0.3.0
google-generativeai>=0.8.0
pandas==2.1.4
python-dotenv==1.0.0
tabulate==0.9.0
